test: move mock sol address checker to new file to share in tests
